Crate panorama_imap

Panorama/IMAP

This is a library that implements the IMAP protocol according to RFC 3501 and several extensions. Although its primary purpose is to be used in panorama, it should be usable for general-purpose IMAP usage. See the client module for more information on how to get started with a client quickly.

Modules

client
IMAP Client

command

parser

response
Structs and enums that have to do with responses.