Struct panorama_imap::client::inner::Client

pub struct Client<C> {
    config: ClientConfig,
    conn: WriteHalf<C>,
    id: usize,
    results: Arc<RwLock<VecDeque<(usize, Option<Response>, Vec<Response>, Option<Waker>)>>>,
    caps: Arc<RwLock<Option<HashSet<Capability>>>>,
    listener_handle: JoinHandle<Result<ReadHalf<C>>>,
    exit_tx: Sender<()>,
    greeting: Arc<RwLock<(Option<Response>, Option<Waker>)>>,
}

The lower-level Client struct, that is shared by all of the exported structs in the state machine.

Fields
config: ClientConfigconn: WriteHalf<C>id: usizeresults: Arc<RwLock<VecDeque<(usize, Option<Response>, Vec<Response>, Option<Waker>)>>>caps: Arc<RwLock<Option<HashSet<Capability>>>>

cached set of capabilities

listener_handle: JoinHandle<Result<ReadHalf<C>>>

join handle for the listener thread

exit_tx: Sender<()>

used for telling the listener thread to stop and return the read half

greeting: Arc<RwLock<(Option<Response>, Option<Waker>)>>

used for receiving the greeting

Implementations

impl<C> Client<C> where
C: AsyncRead + AsyncWrite + Unpin + Send + 'static,

pub fn new(conn: C, config: ClientConfig) -> Self

Creates a new client that wraps a connection

pub fn wait_for_greeting(&self) -> GreetingWaiter

Returns a future that doesn't resolve until we receive a greeting from the server.

pub async fn execute(
&mut self,
cmd: Command
) -> Result<(Response, Vec<Response>)>

Sends a command to the server and returns a handle to retrieve the result

pub async fn capabilities(&mut self, force: bool) -> Result<()>

Executes the CAPABILITY command

pub async fn upgrade(self) -> Result<Client<TlsStream<C>>>

Attempts to upgrade this connection using STARTTLS

pub async fn has_capability(&mut self, cap: impl AsRef<str>) -> Result<bool>

Check if this client has a particular capability