Module panorama_imap::parser::old

Module that implements parsers for all of the IMAP types.

Structs

Rfc3501

Enums

Rule

Constants

_PEST_GRAMMAR_Rfc3501

Functions

build_capabilities

build_capability

build_envelope

build_flag

build_flag_list

build_literal

build_mailbox_data

build_mailbox_list

build_mbx_list_flags

build_msg_att

build_msg_att_static

build_nstring
Wrapper around build_string, except return None for the nil case

build_number
Extracts a numerical type, generic over anything that could possibly be read as a number

build_resp_cond_state

build_resp_text

build_resp_text_code

build_response

build_status

build_string
Extracts a string-type, discarding the surrounding quotes and unescaping the escaped characters

parse_capability

parse_literal

parse_response

unwrap1
Unwraps a singleton pair (a pair that only has one element in its inner list)

Type Definitions

ParseResult